Move file operations to own source files
