+++ cmd/sad/main_test.go
@@ -5,7 +5,6 @@ import (
 	"io/ioutil"
 	"os"
 	"path/filepath"
-	"strings"
 	"testing"
 
 	"github.com/jswny/sad"
@@ -176,65 +175,6 @@ func TestGetRelativeDeploymentFiles(t *testing.T) {
 	if string(content) != string(data) {
 		t.Errorf("Expected file content %s but got %s", content, data)
 	}
-}
-
-func TestFindFilePathRecursive(t *testing.T) {
-	dirName := "dir.test"
-	fileName := "file.test"
-
-	tempDirPath, err := ioutil.TempDir("", dirName)
-
-	if err != nil {
-		t.Fatalf("Error creating temp dir: %s", err)
-	}
-
-	defer os.RemoveAll(tempDirPath)
-
-	tempFile, err := ioutil.TempFile(tempDirPath, fileName)
-
-	if err != nil {
-		t.Fatalf("Error creating temp file: %s", err)
-	}
-
-	tempFileName := filepath.Base(tempFile.Name())
-
-	stringPathSeparator := string(os.PathSeparator)
-	splitTempDirPath := strings.Split(tempDirPath, stringPathSeparator)
-	splitTempDirPathLen := len(splitTempDirPath)
-	mainTempDir := strings.Join(splitTempDirPath[:splitTempDirPathLen-1], stringPathSeparator)
-
-	path, err := main.FindFilePathRecursive(mainTempDir, tempFileName)
-
-	if err != nil {
-		t.Fatalf("Error finding recursive file path: %s", err)
-	}
-
-	expected := tempFile.Name()
-	actual := path
-
-	testutils.CompareStrings(expected, actual, "file path", t)
-}
-
-func TestFindFilePathRecursiveNotFound(t *testing.T) {
-	cwdPath, err := os.Getwd()
-
-	if err != nil {
-		t.Fatalf("Error getting current working directory: %s", err)
-	}
-
-	fileName := "1234556789"
-	path, err := main.FindFilePathRecursive(cwdPath, fileName)
-
-	if err == nil {
-		t.Fatalf("Expected error but got nil!")
-	}
-
-	testutils.CompareStrings(main.FindFilePathRecursiveFileNotFoundErrorMessage, err.Error(), "error", t)
-
-	expected := ""
-	actual := path
-
-	testutils.CompareStrings(expected, actual, "file path", t)
 }
 
 func buildArgs(stringOpts *testutils.StringOptions) []string {
